+++ configserver/src/main/java/com/yahoo/vespa/config/server/rpc/ConfigResponseFactory.java
@@ -1,5 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.rpc;
 
 import com.yahoo.config.codegen.InnerCNode;
 import com.yahoo.vespa.config.ConfigPayload;
@@ -13,6 +13,7 @@ import com.yahoo.vespa.config.protocol.ConfigResponse;
  * @since 5.19
  */
 public interface ConfigResponseFactory {
+
     /**
      * Create a {@link ConfigResponse} for a given payload and generation.
      * @param payload The {@link com.yahoo.vespa.config.ConfigPayload} to put in the response.
@@ -20,4 +21,5 @@ public interface ConfigResponseFactory {
      * @param generation The payload generation.  @return A {@link ConfigResponse} that can be sent to the client.
      */
     ConfigResponse createResponse(ConfigPayload payload, InnerCNode defFile, long generation);
+
 }
