+++ configserver/src/main/java/com/yahoo/vespa/config/server/rpc/LZ4ConfigResponseFactory.java
@@ -1,9 +1,10 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.rpc;
 
 import com.yahoo.config.codegen.InnerCNode;
 import com.yahoo.text.Utf8Array;
 import com.yahoo.vespa.config.ConfigPayload;
+import com.yahoo.vespa.config.LZ4PayloadCompressor;
 import com.yahoo.vespa.config.protocol.CompressionInfo;
 import com.yahoo.vespa.config.protocol.CompressionType;
 import com.yahoo.vespa.config.protocol.ConfigResponse;
@@ -11,17 +12,21 @@ import com.yahoo.vespa.config.protocol.SlimeConfigResponse;
 import com.yahoo.vespa.config.util.ConfigUtils;
 
 /**
- * Simply returns an uncompressed payload.
+ * Compressor that compresses config payloads to lz4.
  *
  * @author lulf
  * @since 5.19
  */
-public class UncompressedConfigResponseFactory implements ConfigResponseFactory {
+public class LZ4ConfigResponseFactory implements ConfigResponseFactory {
+
+    private static LZ4PayloadCompressor compressor = new LZ4PayloadCompressor();
+
     @Override
     public ConfigResponse createResponse(ConfigPayload payload, InnerCNode defFile, long generation) {
         Utf8Array rawPayload = payload.toUtf8Array(true);
         String configMd5 = ConfigUtils.getMd5(rawPayload);
-        CompressionInfo info = CompressionInfo.create(CompressionType.UNCOMPRESSED, rawPayload.getByteLength());
-        return new SlimeConfigResponse(rawPayload, defFile, generation, configMd5, info);
+        CompressionInfo info = CompressionInfo.create(CompressionType.LZ4, rawPayload.getByteLength());
+        Utf8Array compressed = new Utf8Array(compressor.compress(rawPayload.getBytes()));
+        return new SlimeConfigResponse(compressed, defFile, generation, configMd5, info);
     }
 }
