+++ configserver/src/main/resources/configserver-app/services.xml
@@ -6,7 +6,7 @@
     <component id="com.yahoo.vespa.config.server.ConfigServerBootstrap" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.monitoring.Metrics" bundle="configserver" />
     <component id="com.yahoo.vespa.zookeeper.ZooKeeperServer" bundle="zkfacade" />
-    <component id="com.yahoo.vespa.config.server.RpcServer" bundle="configserver" />
+    <component id="com.yahoo.vespa.config.server.rpc.RpcServer" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.ConfigServerDB" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.session.FileDistributionFactory" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.modelfactory.ModelFactoryRegistry" bundle="configserver" />
