+++ configserver/src/test/java/com/yahoo/vespa/config/server/ConfigServerBootstrapTest.java
@@ -5,6 +5,7 @@ import com.yahoo.cloud.config.ConfigserverConfig;
 import com.yahoo.config.provision.TenantName;
 import com.yahoo.io.IOUtils;
 import com.yahoo.vespa.config.server.monitoring.Metrics;
+import com.yahoo.vespa.config.server.rpc.UncompressedConfigResponseFactory;
 import com.yahoo.vespa.config.server.tenant.TenantRequestHandler;
 import com.yahoo.vespa.config.server.tenant.TestWithTenant;
 import com.yahoo.vespa.config.server.version.VersionState;
@@ -71,7 +72,7 @@ public class ConfigServerBootstrapTest extends TestWithTenant {
         }
     }
 
-    public static class MockRpc extends com.yahoo.vespa.config.server.MockRpc {
+    public static class MockRpc extends com.yahoo.vespa.config.server.rpc.MockRpc {
         public volatile boolean started = false;
         public volatile boolean stopped = false;
