+++ configserver/src/test/java/com/yahoo/vespa/config/server/rpc/RpcServerTest.java
@@ -1,5 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.rpc;
 
 import com.google.common.base.Joiner;
 import com.yahoo.cloud.config.LbServicesConfig;
@@ -13,6 +13,8 @@ import com.yahoo.jrt.Request;
 import com.yahoo.vespa.config.*;
 import com.yahoo.vespa.config.protocol.*;
 import com.yahoo.config.provision.ApplicationId;
+import com.yahoo.vespa.config.server.ApplicationSet;
+import com.yahoo.vespa.config.server.ServerCache;
 import com.yahoo.vespa.config.server.application.Application;
 import com.yahoo.vespa.config.server.monitoring.MetricUpdater;
 import com.yahoo.vespa.config.util.ConfigUtils;
